+++ app/src/main/java/com/example/sean/trackmyjob/Models/TimeDiff.kt
@@ -4,6 +4,6 @@ package com.example.sean.trackmyjob.Models
  * @param hours : Hours worked
  * @param minutes : Minutes worked
  */
-data class TimeDiff(internal var hours: Long = 0, internal var minutes: Long = 0)
+data class TimeDiff(internal var hours: Int = 0, internal var minutes: Int = 0)
 {
 }
